fix: standardize parameter descriptions in delete ports and update logging in use cases

+++ src/main/java/com/vluepixel/vetmanager/api/appointment/type/adapter/in/controller/AppointmentTypeController.java
@@ -129,7 +129,7 @@ public final class AppointmentTypeController {
      */
     @Operation(summary = "Delete an appointment type")
     @DeleteMapping("/{id}")
-    public ResponseEntity<BasicResponse> delete(@PathVariable Long id)
+    public ResponseEntity<BasicResponse> delete(@PathVariable Integer id)
             throws NotFoundException {
         return ok(() -> deleteAppointmentTypePort.deleteById(id),
                 "Tipo de cita eliminada",
